Use morepaperlib for runtime command registration on spigot
