+++ bukkit/src/main/java/net/william278/huskhomes/BukkitHuskHomes.java
@@ -65,6 +65,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.TestOnly;
 import space.arim.morepaperlib.MorePaperLib;
+import space.arim.morepaperlib.commands.CommandRegistration;
 import space.arim.morepaperlib.scheduling.GracefulScheduling;
 
 import java.io.File;
@@ -177,11 +178,11 @@ public class BukkitHuskHomes extends JavaPlugin implements HuskHomes, BukkitTask
             }
         });
 
-        // Register events
-        initialize("events", (plugin) -> this.eventListener = new BukkitEventListener(this));
+        // Register event listener
+        initialize("events", (plugin) -> this.eventListener = registerListener());
 
         // Register commands
-        initialize("commands", (plugin) -> this.commands = registerCommands());
+        initialize("commands", (plugin) -> this.commands = BukkitCommand.Type.getCommands(this));
 
         // Initialize the API
         initialize("API", (plugin) -> HuskHomesAPI.register(this));
@@ -191,9 +192,10 @@ public class BukkitHuskHomes extends JavaPlugin implements HuskHomes, BukkitTask
         this.checkForUpdates();
     }
 
+    // Register the event listener
     @NotNull
-    public List<Command> registerCommands() {
-        return BukkitCommand.Type.getCommands(this);
+    protected EventListener registerListener() {
+        return new BukkitEventListener(this);
     }
 
     @Override
@@ -240,15 +242,6 @@ public class BukkitHuskHomes extends JavaPlugin implements HuskHomes, BukkitTask
             audiences = null;
         }
         cancelTasks();
-    }
-
-    /**
-     * Returns the adventure Bukkit audiences.
-     *
-     * @return The adventure Bukkit audiences
-     */
-    public BukkitAudiences getAudiences() {
-        return audiences;
     }
 
     @Override
@@ -483,8 +476,18 @@ public class BukkitHuskHomes extends JavaPlugin implements HuskHomes, BukkitTask
     }
 
     @NotNull
+    public BukkitAudiences getAudiences() {
+        return audiences;
+    }
+
+    @NotNull
     public GracefulScheduling getScheduler() {
         return paperLib.scheduling();
+    }
+
+    @NotNull
+    public CommandRegistration getCommandRegistrar() {
+        return paperLib.commandRegistration();
     }
 
     @Override
